backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.10.3
pydantic-settings>=2.5.2
python-dotenv>=1.0.0
prisma==0.15.0
celery==5.4.0
redis==5.1.1
aiofiles==24.1.0
python-multipart==0.0.12
clerk-backend-api==1.7.0
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
boto3==1.35.0
psycopg2-binary==2.9.9
